pyproject.toml
[project]
name = "mcp-client"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "httpx>=0.28.1",
    "mcp>=1.6.0",
    "mcp-server-time>=0.6.2",
    "openai>=1.70.0",
    "python-dotenv>=1.1.0",
    "watchdog>=6.0.0",
]
